pyproject.toml
[tool.poetry]
name = "videoprocessing"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"

psycopg2-binary = "^2.9.3"
Django = "^4.0.4"
django-rest-framework = "^0.1.0"
boto3 = "^1.22.9"
opencv-python = "^4.5.5"
numpy = "^1.22.3"
celery = "^5.2.6"
redis = "^4.2.2"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"
pylint = "^2.13.8"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
